Draft hardware abstraction layer & refactoring of vent.io (Replaces previous PR of hardware-dev with a cleaner history)
