+++ vent/io/config/devices.ini
@@ -1,0 +1,49 @@
+#Configuration file used to identify specific hardware devices on the Ventilator 
+
+[adc]
+type   = ADS1115
+module = devices
+i2c_address = 0x48
+i2c_bus = 1
+
+[inlet_valve]
+type   = OnOffValve
+module = devices.valves
+pin    = 27
+form   = Normally Closed
+
+[control_valve]
+type   = PWMControlValve
+module = devices.valves
+pin    = 12
+form   = Normally Closed
+#response_curve = path/to/file
+
+[expiratory_valve]
+# Change 'control' to True and change type if you plan to use the expiratory valve for more than on/off
+type    = OnOffValve
+module  = devices.valves
+pin     = 17
+form    = Normally Closed
+
+[pressure_sensor]
+# Note: calibration must convert to units cm(h20)
+
+# The following config is for an Amphenol/All Sensors 20 INCH-G-P4V-MINI
+# Datasheet: https://www.mouser.com/datasheet/2/18/DS-0102_Rev_A-1499183.pdf
+type              = AnalogSensor
+module            = devices.sensors
+adc               = adc
+MUX               = 0
+offset_voltage    = 0.25
+output_span       = 4.0
+conversion_factor = 2.54*20
+
+
+[flow_sensor]
+# Note: SFM3200 is an i2c device and does not need additional calibration
+type   = SFM3200
+module = devices.sensors
+
+
+
