+++ vent/io/config/sim-devices.ini
@@ -1,0 +1,35 @@
+#Configuration file used to identify specific hardware devices on the Ventilator 
+
+[inlet_valve]
+type   = SimOnOffValve
+module = devices.valves
+
+[control_valve]
+type   = SimControlValve
+module = devices.valves
+#response_curve = path/to/file
+
+[expiratory_valve]
+# Change 'control' to True and change type if you plan to use the expiratory valve for more than on/off
+type    = SimOnOffValve
+module  = devices.valves
+
+[pressure_sensor]
+# Note: calibration must convert to units cm(h20)
+
+# The following config is for an Amphenol/All Sensors 20 INCH-G-P4V-MINI
+# Datasheet: https://www.mouser.com/datasheet/2/18/DS-0102_Rev_A-1499183.pdf
+type    = SimSensor
+min     = 0
+max     = 42
+module  = devices.sensors
+
+[flow_sensor]
+# Note: SFM3200 is an i2c device and does not need additional calibration
+type   = SimSensor
+min    = 0
+max    = 120
+module = devices.sensors
+
+
+
